feat(relay): split central and relay
